requirements.txt
Django==2.2
django-crispy-forms==1.8.1
